+++ app/components/chart/daily.js
@@ -39,6 +39,7 @@ const SMBGTooltip = vizComponents.SMBGTooltip;
 const CBGTooltip = vizComponents.CBGTooltip;
 const FoodTooltip = vizComponents.FoodTooltip;
 const ReservoirTooltip = vizComponents.ReservoirTooltip;
+const PhysicalTooltip = vizComponents.PhysicalTooltip;
 
 import Header from './header';
 import Footer from './footer';
@@ -70,6 +71,8 @@ const DailyChart = translate()(class DailyChart extends Component {
     onCarbOut: React.PropTypes.func.isRequired,
     onReservoirHover: React.PropTypes.func.isRequired,
     onReservoirOut: React.PropTypes.func.isRequired,
+    onPhysicalHover: React.PropTypes.func.isRequired,
+    onPhysicalOut: React.PropTypes.func.isRequired,
   };
 
   constructor(props) {
@@ -91,6 +94,8 @@ const DailyChart = translate()(class DailyChart extends Component {
       'onCarbOut',
       'onReservoirHover',
       'onReservoirOut',
+      'onPhysicalHover',
+      'onPhysicalOut',
     ];
 
     this.log = bows('Daily Chart');
@@ -319,6 +324,8 @@ class Daily extends Component {
                 onCarbOut={this.handleCarbOut}
                 onReservoirHover={this.handleReservoirHover}
                 onReservoirOut={this.handleReservoirOut}
+                onPhysicalHover={this.handlePhysicalHover}
+                onPhysicalOut={this.handlePhysicalOut}
                 ref="chart" />
             </div>
           </div>
@@ -393,6 +400,16 @@ class Daily extends Component {
           }}
           side={this.state.hoveredReservoir.side}
           reservoir={this.state.hoveredReservoir.data}
+          bgPrefs={this.props.bgPrefs}
+          timePrefs={this.props.timePrefs}
+        />}
+        {this.state.hoveredPhysical && <PhysicalTooltip
+          position={{
+            top: this.state.hoveredPhysical.top,
+            left: this.state.hoveredPhysical.left
+          }}
+          side={this.state.hoveredPhysical.side}
+          physicalActivity={this.state.hoveredPhysical.data}
           bgPrefs={this.props.bgPrefs}
           timePrefs={this.props.timePrefs}
         />}
@@ -614,6 +631,29 @@ class Daily extends Component {
   handleReservoirOut = () => {
     this.setState({
       hoveredReservoir: false
+    });
+  };
+
+  handlePhysicalHover = physical => {
+    var rect = physical.rect;
+    // range here is -12 to 12
+    var hoursOffset = sundial.dateDifference(physical.data.normalTime, this.state.datetimeLocation, 'h');
+    physical.top = rect.top + (rect.height / 2)
+    if(hoursOffset > 5) {
+      physical.side = 'left';
+      physical.left = rect.left;
+    } else {
+      physical.side = 'right';
+      physical.left = rect.left + rect.width;
+    }
+    this.setState({
+      hoveredPhysical: physical
+    });
+  };
+
+  handlePhysicalOut = () => {
+    this.setState({
+      hoveredPhysical: false
     });
   };