+++ app/core/constants.js
@@ -85,6 +85,7 @@ export const DIABETES_DATA_TYPES = [
   'bolus',
   'wizard',
   'food',
+  'physicalActivity',
 ];
 
 export const MGDL_UNITS = t('mg/dL');